+++ auto_route/lib/src/route/route_config.dart
@@ -1,14 +1,14 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:auto_route/src/utils.dart';
+import 'package:collection/collection.dart';
 import 'package:flutter/cupertino.dart';
-import '../matcher/route_matcher.dart';
 
 typedef TitleBuilder = String Function(BuildContext context, RouteData data);
 
 @immutable
 class AutoRoute {
   final String name;
-  final String _path;
+  final String? _path;
   final bool fullMatch;
   final RouteCollection? _children;
   final String? redirectTo;
@@ -35,7 +35,7 @@ class AutoRoute {
     this.title,
     this.keepHistory = true,
     List<AutoRoute>? children,
-  })  : _path = path ?? toKababCase(name),
+  })  : _path = path,
         _children = children != null ? RouteCollection.from(children) : null;
 
   factory AutoRoute({
@@ -68,7 +68,7 @@ class AutoRoute {
     );
   }
 
-  String get path => _path;
+  String get path => _path!;
 
   bool get hasSubTree => _children != null;
 
@@ -79,6 +79,36 @@ class AutoRoute {
   @override
   String toString() {
     return 'RouteConfig{name: $name}';
+  }
+
+  AutoRoute copyWith({
+    String? name,
+    String? path,
+    bool? fullMatch,
+    String? redirectTo,
+    List<AutoRouteGuard>? guards,
+    bool? usesPathAsKey,
+    Map<String, dynamic>? meta,
+    RouteType? type,
+    bool? fullscreenDialog,
+    bool? maintainState,
+    TitleBuilder? title,
+    bool? keepHistory,
+  }) {
+    return AutoRoute._(
+      name: name ?? this.name,
+      path: path ?? _path,
+      fullMatch: fullMatch ?? this.fullMatch,
+      redirectTo: redirectTo ?? this.redirectTo,
+      guards: guards ?? this.guards,
+      usesPathAsKey: usesPathAsKey ?? this.usesPathAsKey,
+      meta: meta ?? this.meta,
+      type: type ?? this.type,
+      fullscreenDialog: fullscreenDialog ?? this.fullscreenDialog,
+      maintainState: maintainState ?? this.maintainState,
+      title: title ?? this.title,
+      keepHistory: keepHistory ?? this.keepHistory,
+    );
   }
 }
 
@@ -198,4 +228,87 @@ class TestRoute extends AutoRoute {
     super.redirectTo,
     super.fullMatch,
   }) : super._(name: name, path: path);
+}
+
+class RouteCollection {
+  final Map<String, AutoRoute> _routesMap;
+
+  RouteCollection(this._routesMap) : assert(_routesMap.isNotEmpty);
+
+  factory RouteCollection.from(List<AutoRoute> routes, {bool root = false}) {
+    final routesMap = <String, AutoRoute>{};
+    for (var r in routes) {
+      if (r._path != null) {
+        throwIf(
+          !root && r.path.startsWith('/'),
+          'Sub-paths can not start with a "/"',
+        );
+        throwIf(
+          root && !r.path.startsWith(RegExp('[/]|[*]')),
+          'Root-paths must start with a "/" or be a wild-card',
+        );
+        routesMap[r.name] = r;
+      } else {
+        routesMap[r.name] = r.copyWith(
+          path: _generateRoutePath(r.name, root),
+        );
+      }
+    }
+
+    return RouteCollection(routesMap);
+  }
+
+  Iterable<AutoRoute> get routes => _routesMap.values;
+
+  AutoRoute? operator [](String key) => _routesMap[key];
+
+  bool containsKey(String key) => _routesMap.containsKey(key);
+
+  RouteCollection subCollectionOf(String key) {
+    assert(this[key]?.children != null, "$key does not have children");
+    return this[key]!.children!;
+  }
+
+  List<AutoRoute> findPathTo(String routeName) {
+    final track = <AutoRoute>[];
+    for (final route in routes) {
+      if (_findPath(route, routeName, track)) {
+        break;
+      }
+    }
+    return track;
+  }
+
+  bool _findPath(AutoRoute node, String routeName, List<AutoRoute> track) {
+    if (node.name == routeName) {
+      track.add(node);
+      return true;
+    }
+
+    if (node.hasSubTree) {
+      for (AutoRoute child in node.children!.routes) {
+        if (_findPath(child, routeName, track)) {
+          track.insert(0, node);
+          return true;
+        }
+      }
+    }
+
+    return false;
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is RouteCollection &&
+          runtimeType == other.runtimeType &&
+          const MapEquality().equals(_routesMap, other._routesMap);
+
+  @override
+  int get hashCode => const MapEquality().hash(_routesMap);
+
+  static String _generateRoutePath(String name, bool root) {
+    final kebabCased = toKebabCase(name);
+    return root ? '/$kebabCased' : kebabCased;
+  }
 }