+++ auto_route/lib/src/router/controller/auto_router_delegate.dart
@@ -34,9 +34,6 @@ class AutoRouterDelegate extends RouterDelegate<UrlState> with ChangeNotifier {
   @override
   Future<bool> popRoute() async => controller.popTop();
 
-
-
-
   late List<NavigatorObserver> _navigatorObservers;
 
   AutoRouterDelegate(