+++ auto_route/test/simple_router/router.dart
@@ -7,6 +7,6 @@ class SimpleRouter extends MainRouter {
   final List<AutoRoute> routes = [
     AutoRoute(page: FirstRoute.page, path: '/'),
     AutoRoute(page: SecondRoute.page),
-    AutoRoute(page: ThirdRoute.page),
+    AutoRoute(page: ThirdRoute.page,children: [MaterialRoute(page: ThirdRoute.page)]),
   ];
 }